Refactor gatsby-node to separate utils
